Linux System Administration
Step 1: Inspect the file permissions of each of the files below.  This should have already been completed in the activity, but let's double check!
If they do not match the descriptions, please update the permissions. 
Step 2: This step asks you to set up various users.
Step 3: Now we want to execute the commands to fully set up a group on our system. This requires us to create a group, add users to it, create a shared group folder, set the group folder owners for these shared folders. 
Step 4: The final step on your administrator's list involves running an audit against the system in order to harden it.